refactor: package modules for tests and align sidebar inputs
